Create a benchmark specific conftest
